Fixes to align with RFC 6238 algorithm. Added Test Vectors from RFC
The tests make provides assurance that this server can serve any OAUTH
TOTP client as long as the provisioning can be done. The default
implementation expects client apps to provision by connecting directly
to the server. Alignment with apps that use QR code to provision devices
can be achieved by implementing a front end server to provision and
transmit the secret using QR-Code